+++ app/ui/timetable/create-timetable/create-timetable-constants.ts
